backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
groq>=0.9.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
# --- Optional TTS ---
# pyttsx3>=2.90
# gTTS>=2.5.0
# httpx>=0.27.0
